Added `babel-runtime` to the sample project
This is a follow up to an issue reported regarding the usage
of the sample project with `npm@2`, fixed in the template
project.

Under `npm@3` enabled environments, the sample project runs
fine, development and distribution tasks conclude without
problems. But warnings are issued in the terminal about
missing Babel runtime helpers running the sample project under
`npm@2`.  This is due how `npm@3` attempts to organize its
`node_modules` directory as a flat structure, avoiding
packages subdirectories, and `babel-runtime` gets moved to the
top, obscuring the issue.

Adding the latest `babel-runtime` to the project seems
reasonable, but several Babel plugins depend on an outdated
package, so I'm adding version 5.8.34 to avoid breaking the
other plugins.

This patch also upgrades the minimum versions of other Babel
related dependencies.