Update `states.js` when creating a new game state
Now, the `src/states.js` module is updated on the fly, when a
new game state file is created.

The `states.js` file must be declared in the `.yo-rc.json`
configuration, added by Yeoman when creating new projects,
identified by a key named 'states-module'. Projects created
with former versions of the generator (<0.4.0) may take
advantage of this feature too, provided the configuration key
is there.