Revised splash screen scene.
The former implementation relied on some anti-patterns to work. Now,
instead of creating a 'Loader' scene at runtime to load game assets in
parallel, the loading is handled by the splash screen itself, and the
progress bar animation uses features developers are more familiar with.

This commit also removes the `assets.js` module, where static assets
used to be declared, in favor of the regular API. This departs from a
model used since Phaser 2/CE.